+++ app/Propstack/Imports/v1/Objects.php
@@ -142,10 +142,8 @@ class Objects extends Import_Base {
 		$process_handler->set_status( __( 'Import of objects starting', 'connector-for-propstack' ) );
 		$process_handler->set_running( time() );
 
-		// add a log entry if debug is enabled.
-		if ( 1 === absint( get_option( 'propstack_connector_debug', 0 ) ) ) {
-			Log::get_instance()->add( __( 'Import of objects starting', 'connector-for-propstack' ), 'info', 'import' );
-		}
+		// add a log entry.
+		Log::get_instance()->add( __( 'Import of objects has started.', 'connector-for-propstack' ), 'success', 'import' );
 
 		// register a shutdown handler to catch fatal errors during the import.
 		register_shutdown_function( array( $this, 'handle_fatal_shutdown' ) );
@@ -173,11 +171,9 @@ class Objects extends Import_Base {
 
 				// loop through each enabled language and collect its objects.
 				foreach ( $languages as $language_code => $language_enabled ) {
-					// add a log entry if debug is enabled.
-					if ( 1 === absint( get_option( 'propstack_connector_debug', 0 ) ) ) {
-						/* translators: a title will replace %1$s. */
-						Log::get_instance()->add( sprintf( __( 'Import of objects in language %1$s starting', 'connector-for-propstack' ), ' <em>' . $language_code . '</em>' ), 'info', 'import' );
-					}
+					// add a log entry.
+					/* translators: a title will replace %1$s. */
+					Log::get_instance()->add( sprintf( __( 'Import of objects in language %1$s starting', 'connector-for-propstack' ), ' <em>' . $language_code . '</em>' ), 'info', 'import' );
 
 					// load all objects for this language, page by page.
 					$objects = $this->get_objects_from_api( $language_code );
@@ -210,7 +206,7 @@ class Objects extends Import_Base {
 					if ( get_option( 'cfprop_md5_' . $language_code ) === $md5 && 1 !== absint( get_option( 'propstack_connector_debug', 0 ) ) ) {
 						// add a log entry.
 						/* translators: a title will replace %1$s. */
-						Log::get_instance()->add( sprintf( __( 'The objects in the %1$s language have not been modified. The import will not be performed.', 'connector-for-propstack' ), ' <em>' . $language_code . '</em>' ), 'info', 'import' );
+						Log::get_instance()->add( sprintf( __( 'The objects in the %1$s language have not been modified. The import will not be performed.', 'connector-for-propstack' ), ' <em>' . $language_code . '</em>' ), 'success', 'import' );
 
 						/**
 						 * Run actions if objects in Propstack did not change.
@@ -266,10 +262,8 @@ class Objects extends Import_Base {
 				$this->set_max_count( $process_handler, absint( $import_data['total'] ) );
 				$this->set_new_status( $process_handler, __( 'Import of objects is running', 'connector-for-propstack' ) );
 
-				// add a log entry if debug is enabled.
-				if ( 1 === absint( get_option( 'propstack_connector_debug', 0 ) ) ) {
-					Log::get_instance()->add( __( 'Import of objects is running', 'connector-for-propstack' ), 'info', 'import' );
-				}
+				// add a log entry.
+				Log::get_instance()->add( __( 'Import of objects is running', 'connector-for-propstack' ), 'info', 'import' );
 			}
 
 			// get the name of the post-type to use.
@@ -318,11 +312,9 @@ class Objects extends Import_Base {
 						// update the counter.
 						$this->set_count( $process_handler, $process_handler->get_count() + 1 );
 
-						// add a log entry if debug is enabled.
-						if ( 1 === absint( get_option( 'propstack_connector_debug', 0 ) ) ) {
-							/* translators: %1$s will be replaced by the object title. */
-							Log::get_instance()->add( sprintf( __( 'Import of object %1$s prevented.', 'connector-for-propstack' ), '<em>' . $object['title']['value'] . '</em>' ), 'info', 'import' );
-						}
+						// add a log entry.
+						/* translators: %1$s will be replaced by the object title. */
+						Log::get_instance()->add( sprintf( __( 'Import of object %1$s prevented.', 'connector-for-propstack' ), '<em>' . $object['title']['value'] . '</em>' ), 'info', 'import' );
 
 						// update tick.
 						if ( $progress ) {
@@ -336,22 +328,18 @@ class Objects extends Import_Base {
 					/* translators: %1$s will be replaced by the object title. */
 					$this->set_new_status( $process_handler, sprintf( __( 'Import of object %1$s', 'connector-for-propstack' ), '<em>' . $object['title']['value'] . '</em>' ) );
 
-					// add a log entry if debug is enabled.
-					if ( 1 === absint( get_option( 'propstack_connector_debug', 0 ) ) ) {
-						/* translators: %1$s will be replaced by the object title. */
-						Log::get_instance()->add( sprintf( __( 'Import of object %1$s', 'connector-for-propstack' ), '<em>' . $object['title']['value'] . '</em>' ), 'info', 'import' );
-					}
+					// add a log entry.
+					/* translators: %1$s will be replaced by the object title. */
+					Log::get_instance()->add( sprintf( __( 'Import of object %1$s', 'connector-for-propstack' ), '<em>' . $object['title']['value'] . '</em>' ), 'info', 'import' );
 
 					// get the object with the given ID.
 					$propstack_immo_object = ImmoObjects::get_instance()->get_object_by_object_id( $object['id'], $language_code );
 
 					// if the object does not exist, create it.
 					if ( ! $propstack_immo_object instanceof \ConnectorForPropstack\Propstack\ImmoObject ) {
-						// add a log entry if debug is enabled.
-						if ( 1 === absint( get_option( 'propstack_connector_debug', 0 ) ) ) {
-							/* translators: %1$s will be replaced by the object title. */
-							Log::get_instance()->add( sprintf( __( 'Creating new entry for the object %1$s', 'connector-for-propstack' ), '<em>' . $object['title']['value'] . '</em>' ), 'info', 'import' );
-						}
+						// add a log entry.
+						/* translators: %1$s will be replaced by the object title. */
+						Log::get_instance()->add( sprintf( __( 'Creating new entry for the object %1$s', 'connector-for-propstack' ), '<em>' . $object['title']['value'] . '</em>' ), 'info', 'import' );
 
 						// prepare the query to insert a new object.
 						$query = array(
@@ -456,8 +444,8 @@ class Objects extends Import_Base {
 			if ( $progress ) {
 				$progress->finish(); }
 
-			// add a log entry if debug is enabled.
-			if ( Helper::is_development_mode() && 1 === absint( get_option( 'propstack_connector_debug', 0 ) ) ) {
+			// add a log entry.
+			if ( Helper::is_development_mode() ) {
 				Log::get_instance()->add(
 					sprintf(
 						'Memory after chunk: %1$s of %2$s, work list entries: %3$d',
@@ -569,10 +557,8 @@ class Objects extends Import_Base {
 				// log the errors.
 				$this->save_errors_in_log();
 
-				// add a log entry if debug is enabled.
-				if ( 1 === absint( get_option( 'propstack_connector_debug', 0 ) ) ) {
-					Log::get_instance()->add( __( 'Import of objects has been ended.', 'connector-for-propstack' ), 'info', 'import' );
-				}
+				// add a log entry.
+				Log::get_instance()->add( __( 'Import of objects has been ended.', 'connector-for-propstack' ), 'success', 'import' );
 
 				// update the running marker.
 				$process_handler->set_running( 0 );